Inconsistent overridden method
You're changing the method's type while overriding it in the child class. This may lead to unintentional bugs during runtime because of the incinsistency. Some of the cases where this issue would be raised are: - Overriding a method as a property, or the other way around. - Changing a method to async during overriding, or vice versa.